# docs/architecture/openapi.yaml
...
openapi: 3.1.0
info:
  title: API Análisis Predictivo — Desarrollos Comerciales del Sur, S.A.
  description: Contrato del prototipo de tesis. Autenticación JWT y control de acceso RBAC.
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Desarrollo local
paths:
...
  /api/v1/etl/upload-excel:
    post:
      tags:
      - ETL
      summary: Ingesta de histórico de ventas desde Excel
      description: Requiere permiso `etl:cargar`. Valida estructura, tipos y reglas de negocio. Si existen
        errores bloqueantes el lote se rechaza completo (modo estricto) salvo `modo=parcial`. Un archivo
        con checksum ya cargado devuelve 400 `LOTE_DUPLICADO`. Con el arranque cerrado (`GET /etl/config`)
        devuelve 409 `EXCEL_DESHABILITADO`.
      operationId: upload_excel_api_v1_etl_upload_excel_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_excel_api_v1_etl_upload_excel_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EtlResult'
        '422':
          description: El archivo no cumple validaciones estructurales o de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EtlValidationError'
      security:
      - HTTPBearer: []
  /api/v1/etl/batches:
    get:
      tags:
      - ETL
      summary: Historial de lotes ETL
      description: 'Requiere permiso `etl:cargar`. Auditoría de cargas: fecha, filas válidas/rechazadas
        y estado de cada lote, del más reciente al más antiguo.'
      operationId: listar_lotes_api_v1_etl_batches_get
      security:
      - HTTPBearer: []
      parameters:
      - name: estado
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/EstadoLoteEtl'
          - type: 'null'
          title: Estado
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LotePage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/etl/config:
    get:
      tags:
      - ETL
      summary: Política de datos del modelo
      description: Requiere permiso `etl:cargar`. `carga_excel_habilitada=false` indica que el arranque
        está cerrado y las ventas nuevas entran solo por la liquidación diaria; `fuente_reentrenamiento`
        define qué ventas alimentan los reentrenamientos (ADR-14).
      operationId: obtener_config_api_v1_etl_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EtlConfig'
      security:
      - HTTPBearer: []
    put:
      tags:
      - ETL
      summary: Cierra o reabre el arranque y fija la fuente de reentrenamiento
      description: Requiere permiso `etl:configurar` (Administrador). Cambio auditado.
      operationId: actualizar_config_api_v1_etl_config_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EtlConfig'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EtlConfig'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/inventory:
    get:
      tags:
      - Inventory
      summary: Existencias por producto
      description: Requiere permiso `inventario:leer`. `stock_disponible = stock_actual - stock_reservado`.
        Un producto sin fila de inventario se reporta con existencias en 0.
      operationId: listar_existencias_api_v1_inventory_get
      security:
      - HTTPBearer: []
      parameters:
      - name: producto_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Producto Id
      - name: bajo_minimo
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Bajo Minimo
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/inventory/kardex:
    get:
      tags:
      - Inventory
      summary: Movimientos de kardex
      description: 'Requiere permiso `inventario:leer`. Orden: más reciente primero.'
      operationId: listar_kardex_api_v1_inventory_kardex_get
      security:
      - HTTPBearer: []
      parameters:
      - name: producto_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Producto Id
      - name: tipo_movimiento
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/TipoMovimiento'
          - type: 'null'
          title: Tipo Movimiento
      - name: desde
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Desde
      - name: hasta
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Hasta
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KardexPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/inventory/adjustments:
    post:
      tags:
      - Inventory
      summary: Ajuste manual de existencias
      description: 'Requiere `inventario:ajustar`. Tipos: `incremento`, `decremento`, `fijar`. Registra
        un movimiento `ajuste` en el kardex (cantidad con signo) con el motivo. 404 `PRODUCTO_NO_ENCONTRADO`;
        409 `PRODUCTO_INACTIVO`; 400 `STOCK_INSUFICIENTE` si deja el stock bajo lo reservado, `CANTIDAD_INVALIDA`
        o `AJUSTE_SIN_CAMBIO`.'
      operationId: ajustar_existencias_api_v1_inventory_adjustments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AjusteRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AjusteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/catalog/routes:
    get:
      tags:
      - Catalog
      summary: Rutas activas
      description: Cualquier usuario autenticado. Alimenta el filtro de ruta del dashboard.
      operationId: listar_rutas_api_v1_catalog_routes_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RouteItem'
                type: array
                title: Response Listar Rutas Api V1 Catalog Routes Get
      security:
      - HTTPBearer: []
  /api/v1/catalog/categories:
    get:
      tags:
      - Catalog
      summary: Categorías de producto
      description: Cualquier usuario autenticado. Alimenta el formulario de productos.
      operationId: listar_categorias_api_v1_catalog_categories_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CategoriaItem'
                type: array
                title: Response Listar Categorias Api V1 Catalog Categories Get
      security:
      - HTTPBearer: []
  /api/v1/catalog/suppliers:
    get:
      tags:
      - Catalog
      summary: Proveedores activos
      description: Cualquier usuario autenticado. Alimenta el formulario de productos.
      operationId: listar_proveedores_api_v1_catalog_suppliers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProveedorItem'
                type: array
                title: Response Listar Proveedores Api V1 Catalog Suppliers Get
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    AjusteRequest:
      properties:
        producto_id:
          type: string
          format: uuid
          title: Producto Id
        tipo:
          $ref: '#/components/schemas/TipoAjuste'
        cantidad:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*(?:\d{0,10}|(?=[\d.]{1,13}0*$)\d{0,10}\.\d{0,2}0*$)
          title: Cantidad
        motivo:
          type: string
          maxLength: 300
          minLength: 5
          title: Motivo
      type: object
      required:
      - producto_id
      - tipo
      - cantidad
      - motivo
      title: AjusteRequest
      description: 'Ajuste manual (ADR-16). `incremento`/`decremento` exigen `cantidad > 0`; `fijar` establece

        el `stock_actual` resultante (`cantidad >= 0`).'
    AjusteResponse:
      properties:
        movimiento:
          $ref: '#/components/schemas/KardexEntry'
        existencias:
          $ref: '#/components/schemas/StockItem'
      type: object
      required:
      - movimiento
      - existencias
      title: AjusteResponse
...
    Body_upload_excel_api_v1_etl_upload_excel_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        modo:
          $ref: '#/components/schemas/ModoCarga'
          default: estricto
        hoja:
          anyOf:
          - type: string
          - type: 'null'
          title: Hoja
          description: Nombre de la hoja; por defecto la primera
        periodo:
          anyOf:
          - type: string
          - type: 'null'
          title: Periodo
          description: YYYY-MM. Solo formato ancho; por defecto se infiere del nombre del archivo
      type: object
      required:
      - file
      title: Body_upload_excel_api_v1_etl_upload_excel_post
    CategoriaItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        nombre:
          type: string
          title: Nombre
      type: object
      required:
      - id
      - nombre
      title: CategoriaItem
    CodigoErrorEtl:
      type: string
      enum:
      - COLUMNAS_FALTANTES
      - TIPOS_INVALIDOS
      - DATOS_INVALIDOS
      title: CodigoErrorEtl
...
    ErrorFila:
      properties:
        fila:
          type: integer
          title: Fila
        columna:
          type: string
          title: Columna
        valor:
          anyOf:
          - type: string
          - type: 'null'
          title: Valor
        mensaje:
          type: string
          title: Mensaje
        hoja:
          anyOf:
          - type: string
          - type: 'null'
          title: Hoja
      type: object
      required:
      - fila
      - columna
      - mensaje
      title: ErrorFila
      description: Error de validación localizado por fila/columna (fila = número de fila en Excel).
...
    EstadoLoteEtl:
      type: string
      enum:
      - recibido
      - validado
      - rechazado
      - cargado
      title: EstadoLoteEtl
...
    EtlConfig:
      properties:
        carga_excel_habilitada:
          type: boolean
          title: Carga Excel Habilitada
        fuente_reentrenamiento:
          $ref: '#/components/schemas/FuenteReentrenamiento'
      type: object
      required:
      - carga_excel_habilitada
      - fuente_reentrenamiento
      title: EtlConfig
      description: Política de datos del modelo (ADR-14).
    EtlResult:
      properties:
        lote_id:
          type: string
          format: uuid
          title: Lote Id
        estado:
          type: string
          enum:
          - cargado
          - rechazado
          title: Estado
        filas_totales:
          type: integer
          title: Filas Totales
        filas_validas:
          type: integer
          title: Filas Validas
        filas_rechazadas:
          type: integer
          title: Filas Rechazadas
        rango_fechas:
          anyOf:
          - $ref: '#/components/schemas/RangoFechas'
          - type: 'null'
        advertencias:
          items:
            type: string
          type: array
          title: Advertencias
      type: object
      required:
      - lote_id
      - estado
      - filas_totales
      - filas_validas
      - filas_rechazadas
      title: EtlResult
    EtlValidationError:
      properties:
        codigo:
          $ref: '#/components/schemas/CodigoErrorEtl'
        lote_id:
          type: string
          format: uuid
          title: Lote Id
        errores:
          items:
            $ref: '#/components/schemas/ErrorFila'
          type: array
          title: Errores
        total_errores:
          type: integer
          title: Total Errores
      type: object
      required:
      - codigo
      - lote_id
      - errores
      - total_errores
      title: EtlValidationError
...
    FuenteReentrenamiento:
      type: string
      enum:
      - excel_historico
      - excel_mas_liquidacion
      - liquidacion
      title: FuenteReentrenamiento
      description: Qué datos alimentan un entrenamiento (`ml.fuente_reentrenamiento`, ADR-14).
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    KardexEntry:
      properties:
        id:
          type: integer
          title: Id
        producto_id:
          type: string
          format: uuid
          title: Producto Id
        tipo_movimiento:
          $ref: '#/components/schemas/TipoMovimiento'
        cantidad:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Cantidad
        saldo_resultante:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Saldo Resultante
        referencia_tipo:
          anyOf:
          - $ref: '#/components/schemas/ReferenciaTipo'
          - type: 'null'
        referencia_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Referencia Id
        usuario_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Usuario Id
        fecha_movimiento:
          type: string
          format: date-time
          title: Fecha Movimiento
        motivo:
          anyOf:
          - type: string
          - type: 'null'
          title: Motivo
      type: object
      required:
      - id
      - producto_id
      - tipo_movimiento
      - cantidad
      - saldo_resultante
      - fecha_movimiento
      title: KardexEntry
    KardexPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/KardexEntry'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: KardexPage
...
    LoteItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        archivo_nombre:
          type: string
          title: Archivo Nombre
        estado:
          type: string
          title: Estado
        filas_totales:
          type: integer
          title: Filas Totales
        filas_validas:
          type: integer
          title: Filas Validas
        filas_rechazadas:
          type: integer
          title: Filas Rechazadas
        cargado_por:
          type: string
          title: Cargado Por
        creado_en:
          type: string
          format: date-time
          title: Creado En
      type: object
      required:
      - id
      - archivo_nombre
      - estado
      - filas_totales
      - filas_validas
      - filas_rechazadas
      - cargado_por
      - creado_en
      title: LoteItem
      description: Fila del historial de lotes (`etl_lotes`) para la auditoría de cargas.
    LotePage:
      properties:
        total:
          type: integer
          title: Total
        lotes:
          items:
            $ref: '#/components/schemas/LoteItem'
          type: array
          title: Lotes
      type: object
      required:
      - total
      - lotes
      title: LotePage
...
    ModoCarga:
      type: string
      enum:
      - estricto
      - parcial
      title: ModoCarga
...
    ProveedorItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        nombre:
          type: string
          title: Nombre
      type: object
      required:
      - id
      - nombre
      title: ProveedorItem
    RangoFechas:
      properties:
        desde:
          type: string
          format: date
          title: Desde
        hasta:
          type: string
          format: date
          title: Hasta
      type: object
      required:
      - desde
      - hasta
      title: RangoFechas
    ReferenciaTipo:
      type: string
      enum:
      - carga_ruta
      - pedido
      - ajuste
      - etl
      title: ReferenciaTipo
...
    RouteItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        codigo:
          type: string
          title: Codigo
        nombre:
          type: string
          title: Nombre
        zona:
          anyOf:
          - type: string
          - type: 'null'
          title: Zona
      type: object
      required:
      - id
      - codigo
      - nombre
      title: RouteItem
...
    StockItem:
      properties:
        producto_id:
          type: string
          format: uuid
          title: Producto Id
        sku:
          type: string
          title: Sku
        nombre:
          type: string
          title: Nombre
        stock_actual:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Stock Actual
        stock_reservado:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Stock Reservado
        stock_disponible:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Stock Disponible
        stock_minimo:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Stock Minimo
        bajo_minimo:
          type: boolean
          title: Bajo Minimo
      type: object
      required:
      - producto_id
      - sku
      - nombre
      - stock_actual
      - stock_reservado
      - stock_disponible
      - stock_minimo
      - bajo_minimo
      title: StockItem
    StockPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/StockItem'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: StockPage
...
    TipoAjuste:
      type: string
      enum:
      - incremento
      - decremento
      - fijar
      title: TipoAjuste
      description: Modalidad de un ajuste manual de existencias (ADR-16).
...
    TipoMovimiento:
      type: string
      enum:
      - entrada
      - salida
      - ajuste
      - reserva
      - liberacion
      title: TipoMovimiento
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer